Move concrete extractors out of Extractor file
